% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.method.R
\name{check.method}
\alias{check.method}
\title{Creates a \code{method} argument}
\usage{
check.method(method, vars, where, blocks, defaultMethod)
}
\value{
Vector of \code{length(blocks)} element with method names
}
\description{
This helper function creates a valid \code{method} vector. The
\code{method} vector is an argument to the \code{mice} function that
specifies the method for each block.
}
\examples{
make.method(nhanes2)
}
\seealso{
\code{\link{mice}}
}
