% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.mdPattern.R
\name{ds.mdPattern}
\alias{ds.mdPattern}
\title{Model Fitted Values}
\usage{
ds.mdPattern(datasources)
}
\arguments{
\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}

\item{regression}{an object of regression model.}

\item{checks}{a boolean, if TRUE (default) checks that verify elements on the server side
such checks lengthen the run-time so the default is FALSE and one can switch these checks
on (set to TRUE) when faced with some error(s).}
}
\value{
a list of fitted values.
}
\description{
Computes the fitted values from objects returned by modeling functions.
}
\details{
Considering \code{y} as a response variable and x as study variable, the fitted values are the y-values that
would expect for the given x-values according to the best-fitting straight line.
}
\section{Dependencies}{

\code{\link{fittedDS}}
}

\author{
Paula Raissa Costa e Silva
}
