% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{A convenience function for logging into an DataSHIELD cluster}
\usage{
login(cluster = c("recap", "opal-demo"))
}
\arguments{
\item{cluster}{Either \code{"recap"} or \code{"opal-demo"}. For \code{"recap"}
you should have set environmental variables \code{RECAP_DEFAULT},
\code{RECAP_TEST}, \code{RECAP_USER} and \code{RECAP_PW}.}
}
\value{
A list.
}
\description{
This function is called for its side effect to log into the
DataSHIELD cluster.
}
