% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.method.R
\name{make.method}
\alias{make.method}
\title{Creates a \code{method} argument}
\usage{
make.method(where, blocks, vars = NULL, defaultMethod = c("pmm",
  "logreg", "polyreg", "polr"))
}
\value{
Vector of \code{length(blocks)} element with method names
}
\description{
This helper function creates a valid \code{method} vector. The
\code{method} vector is an argument to the \code{mice} function that
specifies the method for each block.
}
\examples{
make.method(nhanes2)
}
\seealso{
\code{\link{mice}}
}
